Updated the sshd-lockdown role so it will play nice with ansible’s --check argument
When running with --check, nothing is actually done on the server, so you can’t rely on registered command results, and the variables (although they will exist) will not contain the dictionary members which command includes when it sets them. So checking for presence of stdout before calling find on it allows these items to simply be “skipped” on a run.